+++ praatio/data_classes/point_tier.py
@@ -48,7 +48,6 @@ def _calculateMinAndMaxTime(entries: Sequence[Point], minT=None, maxT=None):
 
 
 class PointTier(textgrid_tier.TextgridTier):
-
     tierType = POINT_TIER
     entryType = Point
 
@@ -187,7 +186,6 @@ class PointTier(textgrid_tier.TextgridTier):
 
         newEntries: List[Point] = []
         for timestamp, label in self.entries:
-
             newTimestamp = timestamp + offset
             utils.checkIsUndershoot(newTimestamp, self.minTimestamp, errorReporter)
             utils.checkIsOvershoot(newTimestamp, self.maxTimestamp, errorReporter)