+++ praatio/praatio_scripts.py
@@ -57,7 +57,6 @@ def audioSplice(
 
     # Ensure all time points involved in splicing fall on zero crossings
     if alignToZeroCrossing is True:
-
         # Cut the splice segment to zero crossings
         spliceDuration = spliceSegment.duration
         spliceZeroStart = spliceSegment.findNearestZeroCrossing(0)
@@ -177,7 +176,6 @@ def spellCheckEntries(
 
     mispelledEntries = []
     for start, end, label in tier.entries:
-
         # Remove punctuation
         for char in punctuationList:
             label = label.replace(char, " ")
@@ -252,7 +250,6 @@ def splitTierEntries(
 
     # Or insert new entries into existing target tier
     else:
-
         for entry in newEntries:
             targetTier.insertEntry(entry, constants.IntervalCollision.ERROR)